TaskMapped representations CI + FakeVideo with missing frames + vre_collage refactor (uses the reader and supports npz/npy)

See merge request video-representations-extractor/video-representations-extractor!147

+++ test/end_to_end/imgur/run.sh
@@ -4,8 +4,7 @@ export CWD=$(dirname $(readlink -f ${BASH_SOURCE[0]}))
 export VRE_ROOT=$CWD/../../../
 export VID=$CWD/test_video.mp4
 
-test -e $VID || \
-    curl "https://gitlab.com/video-representations-extractor/video-representations-extractor/-/raw/master/resources/test_video.mp4" -o $VID
+test -e $VID || curl "https://gitlab.com/video-representations-extractor/video-representations-extractor/-/raw/master/resources/test_video.mp4" -o $VID
 
 # download imgur upload script
 test -e $CWD/imgur.sh || wget https://raw.githubusercontent.com/tremby/imgur.sh/main/imgur.sh -O $CWD/imgur.sh
@@ -17,13 +16,11 @@ n_frames=$(ffprobe -v error -select_streams v:0 -count_packets -show_entries str
 X=$(shuf -i 1-"$n_frames" -n 1)
 # rm -rf $CWD/test_imgur/
 
-# run VRE
-vre $VID --cfg_path $CWD/cfg.yaml -o $CWD/test_imgur/ --start_frame $X --end_frame $((X+1)) \
+# run VRE & collage
+vre $VID --config_path $CWD/cfg.yaml -o $CWD/test_imgur/ --start_frame $X --end_frame $((X+1)) \
     --output_dir_exists_mode skip_computed
-test "$?" -eq 0 || ( echo "VRE failed"; kill $$ )
-vre_collage $CWD/test_imgur/ --overwrite
+vre_collage $CWD/test_imgur/ -o $CWD/collage --config_path $CWD/cfg.yaml --overwrite
 
-out_img=$CWD/test_imgur/collage/$X.png
+out_img=$CWD/collage/$X.png
 test -f $out_img || ( echo "Image $out_img not found"; kill $$ )
-# xdg-open $out_img
 bash $CWD/imgur.sh $out_img
